requirements.txt
django-cors-headers==2.3.0
django-daterange-filter==1.3.0
djangorestframework==3.9.0
html5lib==1.0.1
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
Pygments==2.3.1
pytz==2018.7
PyPDF2==1.26.0
pytz==2018.7
reportlab==3.5.17
six==1.12.0
webencodings==0.5.1
xhtml2pdf==0.2.3